requirements.txt
streamlit
pandas
numpy
google-generativeai 
python-dotenv